requirements.txt
Django==5.2.8
djangorestframework==3.16.1
python-decouple==3.8
requests==2.32.5
django-cors-headers==4.9.0
whitenoise==6.11.0
gunicorn==23.0.0
